convert formatDate to utils
